Be feature/newitem update (#116)
* #105 add : 아이템 update , service와 controller에 반영

* #105 refactor : 아이템 임시저장 flow 수정

* #105 add : temp/end 임시저장 마무리 api 때 필수값 체크 로직 추가

* #105 add : read dto에 readonly 필드 추가

* #105 add : 아이템 임시저장 수정 이어서 할 시 문서없을 때 오류 수정

* #105 add : put 임시저장 end url 등록

* #105 fix : security config에 등록한 POST 메소드 -> PUT 메소드로 수정

* #105 fix : /item/temp/end 뒤 id 오타 수정

* #105 fix : sharing==false(전용)일 때 cartype입력안되면 에러 던지기

* #105 fix : 전용 속성일 시, 차종 없으면 에러 나게 로직 변경

+++ src/main/java/eci/server/NewItemModule/entity/NewItem.java
@@ -9,6 +9,7 @@ import eci.server.ItemModule.repository.color.ColorRepository;
 import eci.server.ItemModule.repository.item.ItemMakerRepository;
 import eci.server.ItemModule.repository.item.ItemTypesRepository;
 import eci.server.ItemModule.repository.member.MemberRepository;
+import eci.server.NewItemModule.dto.newItem.create.NewItemTemporaryContinueRequest;
 import eci.server.NewItemModule.dto.newItem.update.NewItemUpdateRequest;
 import eci.server.NewItemModule.entity.supplier.Maker;
 import eci.server.ItemModule.entity.member.Member;
@@ -17,15 +18,16 @@ import eci.server.NewItemModule.entity.coating.CoatingType;
 import eci.server.NewItemModule.entity.coating.CoatingWay;
 import eci.server.NewItemModule.entity.maker.NewItemMaker;
 import eci.server.NewItemModule.entity.supplier.Supplier;
-import eci.server.NewItemModule.exception.CoatingNotFoundException;
-import eci.server.NewItemModule.exception.SupplierNotFoundException;
+import eci.server.NewItemModule.exception.*;
 import eci.server.NewItemModule.repository.coatingType.CoatingTypeRepository;
 import eci.server.NewItemModule.repository.coatingWay.CoatingWayRepository;
 import eci.server.NewItemModule.repository.maker.NewItemMakerRepository;
 import eci.server.NewItemModule.repository.supplier.SupplierRepository;
 import eci.server.ProjectModule.entity.project.CarType;
 import eci.server.ProjectModule.entity.project.ClientOrganization;
+import eci.server.ProjectModule.exception.CarTypeNotFoundException;
 import eci.server.ProjectModule.exception.ClientOrganizationNotFoundException;
+import eci.server.ProjectModule.repository.carType.CarTypeRepository;
 import eci.server.ProjectModule.repository.clientOrg.ClientOrganizationRepository;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -566,8 +568,8 @@ public class NewItem extends EntityDate {
         added.stream().forEach(i -> {
             attachments.add(i);
             i.initNewItem(this);
-            i.setAttach_comment(req.getAddedAttachmentComment().get((added.indexOf(i))));
-            i.setTag(req.getAddedTag().get((added.indexOf(i))));
+            i.setAttach_comment(req.getAddedAttachmentComment().get((req.getAddedAttachments().indexOf(i))));
+            i.setTag(req.getAddedTag().get((req.getAddedAttachments().indexOf(i))));
             i.setAttachmentaddress(
                     "src/main/prodmedia/image/" +
                             sdf1.format(now).substring(0,10)
@@ -711,6 +713,16 @@ public class NewItem extends EntityDate {
         this.readonly = true;
     }
 
+
+
+    @Getter
+    @AllArgsConstructor
+    public static class NewItemFileUpdatedResult {
+        private NewItemAttachmentUpdatedResult attachmentUpdatedResult;
+        private NewItemImageUpdatedResult imageUpdatedResult;
+    }
+
+
     /**
      * postupdaterequest 받아서 update 수행
      *
@@ -726,7 +738,8 @@ public class NewItem extends EntityDate {
             NewItemMakerRepository itemMakerRepository,
             ItemTypesRepository itemTypesRepository,
             CoatingWayRepository coatingWayRepository,
-            CoatingTypeRepository coatingTypeRepository
+            CoatingTypeRepository coatingTypeRepository,
+            CarTypeRepository carTypeRepository
     ) {
         AtomicInteger k = new AtomicInteger();
 
@@ -738,6 +751,21 @@ public class NewItem extends EntityDate {
                 itemTypesRepository.findById(req.getTypeId()).orElseThrow(ItemNotFoundException::new);
 
         this.sharing = req.isSharing();
+
+        this.carType =
+                this.carType =                 //전용일 때야 차종 생성
+                        (!req.isSharing())?
+                                //1. 전용이라면
+                                req.getCarTypeId()==null?
+                                        //1-1 : 아이디 없으면 (무조건 에러 튕기도록
+                                        carTypeRepository.findById(0L).orElseThrow(CarTypeNotFoundException::new):
+                                        //null 아니면 입력받은 것
+                                        carTypeRepository.findById(req.getCarTypeId()).orElseThrow(CarTypeNotFoundException::new)
+
+                                :
+                                //2. 공용이라면
+                                carTypeRepository.findById(99999L).orElseThrow(CarTypeNotFoundException::new);
+
 
         this.integrate = req.getIntegrate().isBlank()?this.integrate:req.getIntegrate();
 
@@ -753,7 +781,7 @@ public class NewItem extends EntityDate {
 
         this.importance = req.getImportance().isBlank()?this.importance:req.getImportance();
 
-        this.color = req.getColorId()==null?this.color:colorRepository.findById(Long.valueOf(req.getColorId()))
+        this.color = req.getColorId()==null?this.color:colorRepository.findById(req.getColorId())
                 .orElseThrow(ColorNotFoundException::new);
 
         this.loadQuantity = req.getLoadQuantity().isBlank()?this.loadQuantity:req.getLoadQuantity();
@@ -817,9 +845,14 @@ public class NewItem extends EntityDate {
                         req.getAddedAttachments(),
                         req.getDeletedAttachments()
                 );
-        addUpdatedAttachments(req, resultAttachment.getAddedAttachments());
 
-        deleteAttachments(resultAttachment.getDeletedAttachments());
+        if(req.getAddedTag().size()>0) {
+            addUpdatedAttachments(req, resultAttachment.getAddedAttachments());
+        }
+
+        if(req.getDeletedAttachments().size()>0) {
+            deleteAttachments(resultAttachment.getDeletedAttachments());
+        }
 
         NewItemFileUpdatedResult fileUpdatedResult =
                 new NewItemFileUpdatedResult(resultAttachment,resultImage);
@@ -833,12 +866,155 @@ public class NewItem extends EntityDate {
     }
 
 
-    @Getter
-    @AllArgsConstructor
-    public static class NewItemFileUpdatedResult {
-        private NewItemAttachmentUpdatedResult attachmentUpdatedResult;
-        private NewItemImageUpdatedResult imageUpdatedResult;
-    }
+    public NewItemFileUpdatedResult tempEnd(
+            NewItemUpdateRequest req,
+            ColorRepository colorRepository,
+            MemberRepository memberRepository,
+            ClientOrganizationRepository clientOrganizationRepository,
+            SupplierRepository supplierRepository,
+            NewItemMakerRepository itemMakerRepository,
+            ItemTypesRepository itemTypesRepository,
+            CoatingWayRepository coatingWayRepository,
+            CoatingTypeRepository coatingTypeRepository,
+            CarTypeRepository carTypeRepository
+    ) {
 
+        if(req.getClassification1Id()==null || req.getClassification2Id() ==null || req.getClassification3Id()==null){
+            throw new ClassificationRequiredException();
+        }
+        if(req.getClassification1Id()==99999L || req.getClassification2Id() == 99999L || req.getClassification3Id()== 99999L){
+            throw new ProperClassificationRequiredException();
+        }
+        //아이템 타입 체크
+        if(req.getTypeId()==null){
+            throw new ItemTypeRequiredException();
+        }
+
+        //아이템 이름 체크
+        if(req.getName().isBlank()){
+            throw new ItemNameRequiredException();
+        }
+        AtomicInteger k = new AtomicInteger();
+
+        //TODO update할 때 사용자가 기존 값 없애고 보낼 수도 있자나 => fix needed
+        //isBlank 랑 isNull로 판단해서 기존 값 / req 값 채워넣기
+        this.name = req.getName().isBlank()?this.name:req.getName();
+
+        this.itemTypes = req.getTypeId()==null?this.itemTypes:
+                itemTypesRepository.findById(req.getTypeId()).orElseThrow(ItemNotFoundException::new);
+
+        this.sharing = req.isSharing();
+
+        this.carType =                 //전용일 때야 차종 생성
+                (!req.isSharing())?
+                        //1. 전용이라면
+                        req.getCarTypeId()==null?
+                                //1-1 : 아이디 없으면 (무조건 에러 튕기도록
+                                carTypeRepository.findById(0L).orElseThrow(CarTypeNotFoundException::new):
+                                //null 아니면 입력받은 것
+                                carTypeRepository.findById(req.getCarTypeId()).orElseThrow(CarTypeNotFoundException::new)
+
+                        :
+                        //2. 공용이라면
+                        carTypeRepository.findById(99999L).orElseThrow(CarTypeNotFoundException::new);
+
+
+                this.integrate = req.getIntegrate().isBlank()?this.integrate:req.getIntegrate();
+
+        this.curve = req.getCurve().isBlank()?this.curve:req.getCurve();
+
+        this.width = req.getWidth().isBlank()?this.width:req.getWidth();
+
+        this.height= req.getHeight().isBlank()?this.height:req.getHeight();
+
+        this.thickness = req.getThickness().isBlank()?this.thickness:req.getThickness();
+
+        this.weight = req.getWeight().isBlank()?this.weight:req.getWeight();
+
+        this.importance = req.getImportance().isBlank()?this.importance:req.getImportance();
+
+        this.color = req.getColorId()==null?this.color:colorRepository.findById(req.getColorId())
+                .orElseThrow(ColorNotFoundException::new);
+
+        this.loadQuantity = req.getLoadQuantity().isBlank()?this.loadQuantity:req.getLoadQuantity();
+
+        this.forming = req.getForming().isBlank()?this.forming:req.getForming();
+
+        this.coatingWay = req.getCoatingWayId()==null?this.coatingWay:
+                coatingWayRepository.findById
+                        (req.getCoatingWayId()).orElseThrow(CoatingNotFoundException::new);
+
+        this.coatingType = req.getCoatingTypeId()==null?this.coatingType:
+                coatingTypeRepository.findById
+                        (req.getCarTypeId()).orElseThrow(CoatingNotFoundException::new);
+
+
+        this.modulus = req.getModulus().isBlank()?this.modulus:req.getModulus();
+
+        this.screw = req.getScrew().isBlank()?this.screw:req.getScrew();
+
+        this.cuttingType = req.getCuttingType().isBlank()?this.cuttingType:req.getCuttingType();
+
+        this.lcd = req.getLcd().isBlank()?this.lcd:req.getLcd();
+
+        this.displaySize = req.getDisplaySize().isBlank()?this.displaySize:req.getDisplaySize();
+
+        this.screwHeight = req.getScrewHeight().isBlank()?this.screwHeight:req.getScrewHeight();
+
+        this.clientOrganization = req.getClientOrganizationId()==null?this.clientOrganization:
+                clientOrganizationRepository.findById(req.getClientOrganizationId())
+                        .orElseThrow(ClientOrganizationNotFoundException::new);
+
+        this.supplierOrganization = req.getSupplierOrganizationId()==null?this.supplierOrganization:
+                supplierRepository.findById(req.getSupplierOrganizationId())
+                        .orElseThrow(SupplierNotFoundException::new);
+
+//        this.makers =
+//                makers.stream().map(
+//
+//                                //다대다 관계를 만드는 구간
+//                                r -> new NewItemMaker(
+//                                        this, r, partnumbers.get(makers.indexOf(r))
+//                                )
+//                        )
+//                        .collect(toList());
+
+        this.tempsave = true;
+        this.readonly = true; //0605- 이 부분하나가 변경, 이 것은 얘를 false 에서 true로 변경 !
+
+        NewItemImageUpdatedResult resultImage =
+                findImageUpdatedResult(
+                        req.getAddedImages(),
+                        req.getDeletedImages()
+                );
+
+        addImages(resultImage.getAddedImages());
+        deleteImages(resultImage.getDeletedImages());
+
+        NewItemAttachmentUpdatedResult resultAttachment =
+
+                findAttachmentUpdatedResult(
+                        req.getAddedAttachments(),
+                        req.getDeletedAttachments()
+                );
+
+        if(req.getAddedTag().size()>0) {
+            addUpdatedAttachments(req, resultAttachment.getAddedAttachments());
+        }
+
+        if(req.getDeletedAttachments().size()>0) {
+            deleteAttachments(resultAttachment.getDeletedAttachments());
+        }
+
+        NewItemFileUpdatedResult fileUpdatedResult =
+                new NewItemFileUpdatedResult(resultAttachment,resultImage);
+
+        this.modifier =
+                memberRepository.findById(
+                        req.getModifierId()
+                ).orElseThrow(MemberNotFoundException::new);//05 -22 생성자 추가
+
+        return fileUpdatedResult;
+    }
 
 }
